refactor: restructure project exports and module organization
